Lazy load resource drawer
