+++ frontends/main/src/page-components/LearningResourceDrawer/useResourceDrawerHref.ts
@@ -1,0 +1,26 @@
+import { useCallback } from "react"
+import { RESOURCE_DRAWER_QUERY_PARAM } from "@/common/urls"
+import { ReadonlyURLSearchParams, useSearchParams } from "next/navigation"
+
+const getOpenDrawerSearchParams = (
+  current: ReadonlyURLSearchParams,
+  resourceId: number,
+) => {
+  const newSearchParams = new URLSearchParams(current)
+  newSearchParams.set(RESOURCE_DRAWER_QUERY_PARAM, resourceId.toString())
+  return newSearchParams
+}
+
+const useResourceDrawerHref = () => {
+  const searchParams = useSearchParams()
+
+  return useCallback(
+    (resourceId: number) => {
+      const hash = typeof window !== "undefined" && window?.location.hash
+      return `?${getOpenDrawerSearchParams(searchParams, resourceId)}${hash || ""}`
+    },
+    [searchParams],
+  )
+}
+
+export { useResourceDrawerHref }
